kinopark-backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.1
attrs==25.4.0
colorama==0.4.6
decorator==5.2.1
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.1
django-unfold==0.68.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.29.0
executing==2.2.1
filelock==3.18.0
inflection==0.5.1
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
matplotlib-inline==0.2.1
parso==0.8.5
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pure_eval==0.2.3
pygame==2.5.2
Pygments==2.19.2
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
virtualenv==20.29.3
wcwidth==0.2.14
